% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_geography.R
\name{filter_geography}
\alias{filter_geography}
\title{filter_geography}
\usage{
filter_geography(
  x,
  y,
  filter_method,
  min_area_of_intersection = NULL,
  min_distance_from_geography = NULL
)
}
\arguments{
\item{x}{the shapefile to filter}

\item{y}{the shapefile to be used for filtering x}

\item{filter_method}{type of filter to be done ("intersection", "crop", or "distance")}

\item{min_area_of_intersection}{if filter_method = "intersection", the minimum area of intersection between a geometry from x and y}

\item{min_distance_from_geography}{if filter_method = "distance", the minimum distance from a geometry in x to y}
}
\value{
a filtered shapefile
}
\description{
filters a shapefile using the boundary(ies) of another shapefile; utilizes the sf() package 'join' functions.
}
\examples{
filter_geography(x=tract_map, y=city_map, filter_method="intersection", min_area_of_intersection=0)
}
