% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ACS_yearly_pop.R
\name{generate_ACS_yearly_pop}
\alias{generate_ACS_yearly_pop}
\title{generate_ACS_yearly_pop}
\usage{
generate_ACS_yearly_pop(
  census_key,
  state_abr,
  geo_level = "place",
  years_to_include = seq(2014, 2023),
  filter_geos = NULL,
  acs_table = "acs5"
)
}
\arguments{
\item{census_key}{personal Census API key}

\item{state_abr}{state abbreviation(s)}

\item{geo_level}{level of geography the populations should be on; this function can only do "place" level see generate_ACS_yearly_pop_tract for tract-level}

\item{years_to_include}{a vector of years to include}

\item{filter_geos}{a vector of ID's to filter the geography to}

\item{acs_table}{version of ACS to use; default is "acs5"}
}
\value{
data frame; population by polygon
}
\description{
gets yearly ACS populations for census places across multiple years; optionally across multiple states
Note: this was an attempt to generalize generate_ACS_yearly_pop_tract
}
\examples{
generate_ACS_yearly_pop(census_key = census_api_key,
state_abr = unique(citylist$state_abr),
geo_level = "place",
years_to_include = earliest_murder_year:current_year,
filter_geos = str_pad(citylist$stpl_fips, width=7, side="left", pad="0"))

}
