% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ACS_yearly_pop_tract.R
\name{generate_ACS_yearly_pop_tract}
\alias{generate_ACS_yearly_pop_tract}
\title{generate_ACS_yearly_pop_tract}
\usage{
generate_ACS_yearly_pop_tract(
  census_key,
  state_abr,
  years_to_include = seq(2014, 2023),
  filter_tracts = NULL
)
}
\arguments{
\item{census_key}{personal Census API key}

\item{state_abr}{state abbreviation(s)}

\item{years_to_include}{a vector of years to include}

\item{filter_tracts}{a vector of tract ID's to filter the populations to}
}
\value{
data frame; population by tract
}
\description{
gets yearly ACS populations for census tracts across multiple years; this function will impute population values for years not yet available using the average change in prior years
}
\examples{
generate_ACS_yearly_pop_tract(census_key = API_key, state_abr = unique(df_agg$state_abr), years_to_include = unique(df_agg$year), filter_tracts = city_tract_geography$tract_id)

}
