% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_addresses.R
\name{geocode_addresses}
\alias{geocode_addresses}
\title{geocode_addresses}
\usage{
geocode_addresses(API_key, addresses)
}
\arguments{
\item{API_key}{personal Google API key}

\item{addresses}{list of addresses to geocode; vector}
}
\value{
data frame with Google Geocoding output
}
\description{
A wrapper function that converts a list of addresses to lat/long coordinates
}
\examples{
geocode_addresses(API_key = API_key, addresses = df_to_geocode$location)

}
