% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_monthly_population.R
\name{impute_monthly_population}
\alias{impute_monthly_population}
\title{impute_monthly_population}
\usage{
impute_monthly_population(pops_by_year_df, id_col, population_col)
}
\arguments{
\item{pops_by_year_df}{data frame with population values by year and a 'year' column}

\item{id_col}{column names to group by when imputing values (e.g. an ID column)}

\item{population_col}{the names of column(s) to impute}
}
\value{
data frame; population by month and year
}
\description{
imputes population values for months in between years of population data.
Uses average change in population per month to impute and extrapolate population values.
}
\examples{
impute_monthly_population(pops_by_year_df = yearly_pop_df, id_col = c("stpl_fips"), population_col = "place_population_est")

}
