% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_yearly_population.R
\name{impute_yearly_population}
\alias{impute_yearly_population}
\title{impute_yearly_population}
\usage{
impute_yearly_population(
  pop_df,
  group_cols = c("stpl_fips"),
  years_to_impute = min(pop_df$year):max(years_to_impute),
  window = 2
)
}
\arguments{
\item{pop_df}{data frame with population values in a 'place_population_est_year' column and a 'year' column}

\item{group_cols}{column names to group by when imputing values (e.g. an ID column)}

\item{years_to_impute}{all years to include in the final data}

\item{window}{window of rows to average in order to calculate an average change}
}
\value{
data frame with population values by year
}
\description{
imputes population values for years in between census years and for years with population data not yet available.
Uses average change in population per year to impute and extrapolate population values.
}
\examples{
impute_yearly_population(pop_df = pop_df, group_cols = c("stpl_fips"), years_to_impute = min(pop_df$year):max(pop_df$year), window=2)

}
