% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_geography.R
\name{load_geography}
\alias{load_geography}
\title{load_geography}
\usage{
load_geography(
  geography_level,
  census_year = NULL,
  state_abr = NULL,
  counties = NULL,
  fips_code = NULL,
  crs_code = 4326
)
}
\arguments{
\item{geography_level}{geography type to return; "place", "tract", "block", or "block group"}

\item{census_year}{the year of the data to get}

\item{state_abr}{state abbreviation(s) of the geographies to be returned}

\item{counties}{county(ies) of the geographies to be returned}

\item{fips_code}{if geography_level = "place", optionally provide the FIPS code of a census place}

\item{crs_code}{the CRS code of the geography}
}
\value{
sf data frame
}
\description{
Get the geography for census place(s), tract(s), block(s), or block group(s)
}
\examples{
load_geography(geography_level="place", census_year=2020, state_abr="KY", fips_code="2148000")

}
