% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coordinates_to_polygon.R
\name{map_coordinates_to_polygon}
\alias{map_coordinates_to_polygon}
\title{map_coordinates_to_polygon}
\usage{
map_coordinates_to_polygon(
  coordinate_df,
  polygon_map,
  crs_code = 4326,
  longitude_col = "long_ggl",
  latitude_col = "lat_ggl"
)
}
\arguments{
\item{coordinate_df}{data frame with latitude/longitude coordinates}

\item{polygon_map}{the polygons that the coordinates should be mapped to; sf data frame}

\item{crs_code}{CRS code of the geographies}

\item{longitude_col}{name of the longitude column in coordinate_df}

\item{latitude_col}{name of the latitude column in coordinate_df}
}
\value{
sf dataframe; coordinates with the ID column from polygon map
}
\description{
joins a list of coordinates to polygons to determine which polygons each coordinate is in
}
\examples{
map_coordinates_to_polygon(coordinate_df = df \%>\% filter(!is.na(lat_ggl)), polygon_map = city_tract_geography)

}
