% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_place_fips_code.R
\name{search_place_fips_code}
\alias{search_place_fips_code}
\title{search_place_fips_code}
\usage{
search_place_fips_code(city, state_abr)
}
\arguments{
\item{city}{name of city in title case}

\item{state_abr}{state abbreviation that the city is in}
}
\value{
FIPS code for a city
}
\description{
searches for the census place FIPS code given the name and state abbreviation for a city
}
\examples{
search_place_fips_code(city="Chicago", state_abr="IL")

}
